refactor: ReqeustUtil을 Util로 리네이밍
