+++ tomcat/src/main/java/org/apache/coyote/http11/handler/DefaultHandler.java
@@ -5,9 +5,9 @@ import org.apache.coyote.model.request.HttpRequest;
 import org.apache.coyote.model.response.HttpResponse;
 import org.apache.coyote.model.response.ResponseLine;
 import org.apache.coyote.model.response.StatusCode;
-import org.apache.coyote.utils.RequestUtil;
+import org.apache.coyote.utils.Util;
 
-import static org.apache.coyote.utils.RequestUtil.getExtension;
+import static org.apache.coyote.utils.Util.getExtension;
 
 public class DefaultHandler implements Handler {
 
@@ -21,7 +21,7 @@ public class DefaultHandler implements Handler {
     public String getResponse() {
         final String path = httpRequest.getPath();
         final String extension = ContentType.getType(getExtension(path));
-        final String body = RequestUtil.getResponseBody(path, this.getClass());
+        final String body = Util.getResponseBody(path, this.getClass());
 
         return HttpResponse.of(extension, body, ResponseLine.of(StatusCode.OK))
                 .getResponse();
