+++ tomcat/src/main/java/org/apache/coyote/http11/handler/IndexHandler.java
@@ -4,7 +4,7 @@ import org.apache.coyote.model.request.HttpRequest;
 import org.apache.coyote.model.response.HttpResponse;
 import org.apache.coyote.model.response.ResponseLine;
 import org.apache.coyote.model.response.StatusCode;
-import org.apache.coyote.utils.RequestUtil;
+import org.apache.coyote.utils.Util;
 
 import static org.apache.coyote.model.request.ContentType.HTML;
 
@@ -19,7 +19,7 @@ public class IndexHandler implements Handler {
 
     @Override
     public String getResponse() {
-        String responseBody = RequestUtil.getResponseBody(INDEX_HTML, getClass());
+        String responseBody = Util.getResponseBody(INDEX_HTML, getClass());
         ResponseLine responseLine = ResponseLine.of(StatusCode.OK);
         HttpResponse httpResponse = HttpResponse.of(HTML.getExtension(), responseBody, responseLine);
         return httpResponse.getResponse();
