+++ tomcat/src/test/java/org/apache/coyote/utils/UtilTest.java
@@ -10,7 +10,7 @@ import java.util.Map;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class RequestUtilTest {
+class UtilTest {
 
     @Test
     @DisplayName("파일이름으로 파일내용을 읽는다.")
@@ -19,7 +19,7 @@ class RequestUtilTest {
         String filename = "/index.html";
 
         // when
-        String actual = RequestUtil.getResponseBody(filename, this.getClass());
+        String actual = Util.getResponseBody(filename, this.getClass());
 
         // then
         assertThat(actual).isNotBlank();
@@ -32,7 +32,7 @@ class RequestUtilTest {
         String filename = "/nonFile";
 
         // when & then
-        assertThatThrownBy(() -> RequestUtil.getResponseBody(filename, this.getClass()))
+        assertThatThrownBy(() -> Util.getResponseBody(filename, this.getClass()))
                 .isInstanceOf(NotFoundFileException.class);
     }
 
@@ -43,7 +43,7 @@ class RequestUtilTest {
         String filename = "/index.html";
 
         // when
-        String actual = RequestUtil.getExtension(filename);
+        String actual = Util.getExtension(filename);
 
         // then
         assertThat(actual).isEqualTo("html");
@@ -56,7 +56,7 @@ class RequestUtilTest {
         String filename = "/login?account=gugu&password=password";
 
         // when
-        String actual = RequestUtil.calculatePath(filename);
+        String actual = Util.calculatePath(filename);
 
         // then
         assertThat(actual).isEqualTo("/login.html");
@@ -72,7 +72,7 @@ class RequestUtilTest {
         expected.put("password", "password");
 
         // when
-        Map<String, String> actual = RequestUtil.getParam(filename);
+        Map<String, String> actual = Util.getParam(filename);
 
         // then
         assertThat(actual).isEqualTo(expected);
